requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-core
langchain-openai
pymysql
aiomysql
pymilvus
python-dotenv
aiohttp
requests
volcengine
